backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pydantic
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-decouple
redis
aioredis
httpx
jinja2
emails
razorpay
boto3
openai
fpdf2
Pillow
PyMuPDF
slowapi
python-json-logger
pytest
pytest-asyncio
pytest-mock
factory-boy
faker
google-auth 
google-auth-oauthlib 
google-api-python-client
setuptools